requirements.txt
pika
pydantic
psycopg[binary,pool]
python-dotenv
pyyaml